Restfull api: Adds jetty server
This commits adds jetty server that expose 8080 as http server.
The api provides one get request that allows user to ask for commits
in addition, allow user to provide query parameters for page and size
by given owner and repository.

Furthermore, a cache system was added into GitCommitAccess class
to help reduce github access

Finally, a dockerfile was added so we can run in a docker container.